requirements.txt
Django==2.1.7
django-rest-framework==0.1.0
djangorestframework==3.9.1
psycopg2-binary==2.7.7
pytz==2018.9
